# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  Agent:
    properties:
      description:
        type: string
      id:
        type: string
      instructions:
        type: string
      llms:
        items:
          $ref: '#/definitions/AgentToLLM'
        type: array
      name:
        type: string
    type: object
...
  AgentToLLM:
    properties:
      llm:
        $ref: '#/definitions/LLM'
      llm_id:
        type: string
      priority:
        type: integer
    type: object
...
  Call:
    properties:
      arguments:
        additionalProperties: {}
        type: object
      id:
        type: string
      tool:
        type: string
    type: object
...
  CreateLLMDTO:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      name:
        maxLength: 255
        type: string
      provider:
        maxLength: 255
        type: string
    required:
    - configuration
    - name
    - provider
    type: object
  CreateSessionDTO:
    properties:
      agent_id:
        type: string
      metadata:
        $ref: '#/definitions/JSONMap'
    required:
    - agent_id
    type: object
  DriverInfo:
    properties:
      configuration_schema:
        $ref: '#/definitions/JSONMap'
      description:
        type: string
      icon:
        type: string
      key:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - description
    - name
    type: object
...
  Error:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
      statusCode:
        type: integer
    type: object
  ExecutionMessage:
    properties:
      content:
        type: string
      execution_id:
        type: string
      final:
        type: boolean
      id:
        type: string
      role:
        $ref: '#/definitions/Role'
      sequence:
        type: integer
      timestamp:
        type: string
      tool_call_id:
        description: ToolCallID links a RoleTool message back to the tool.Call.ID
          it answers.
        type: string
      tool_calls:
        description: ToolCalls is set on assistant messages that requested tool calls.
        items:
          $ref: '#/definitions/Call'
        type: array
    type: object
...
  JSONMap:
    additionalProperties: {}
    type: object
...
  LLM:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
  LLMResponse:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
  LLMsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/LLM'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
...
  Role:
    enum:
    - system
    - user
    - assistant
    - tool
    type: string
    x-enum-varnames:
    - RoleSystem
    - RoleUser
    - RoleAssistant
    - RoleTool
...
  RunSessionDTO:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  Session:
    properties:
      agent:
        $ref: '#/definitions/Agent'
      agent_id:
        type: string
      client_id:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/JSONMap'
    type: object
  SessionMessagesPage:
    properties:
      data:
        items:
          $ref: '#/definitions/ExecutionMessage'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
  SessionResponse:
    properties:
      agent:
        $ref: '#/definitions/Agent'
      agent_id:
        type: string
      client_id:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/JSONMap'
    type: object
  SessionsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/Session'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
...
  UpdateLLMDTO:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      name:
        maxLength: 255
        type: string
      provider:
        maxLength: 255
        type: string
    type: object
  UpdateSessionDTO:
    properties:
      agent_id:
        type: string
      metadata:
        $ref: '#/definitions/JSONMap'
    type: object
...
info:
  contact: {}
  description: API for the Snipet platform.
  title: Snipet API
  version: "1.0"
paths:
...
  /client/{client_code}/session:
    get:
      description: Lists sessions of a client, with optional pagination and includes.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      - collectionFormat: csv
        description: Related resources to include
        in: query
        items:
          enum:
          - agent
          type: string
        name: include
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SessionsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List sessions
      tags:
      - session
    post:
      consumes:
      - application/json
      description: Creates a new session for a client.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateSessionDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/SessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create session
      tags:
      - session
  /client/{client_code}/session/{id}:
    delete:
      description: Deletes a session by ID.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete session
      tags:
      - session
    get:
      description: Returns a session by ID.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - collectionFormat: csv
        description: Related resources to include
        in: query
        items:
          enum:
          - agent
          type: string
        name: include
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SessionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get session
      tags:
      - session
    put:
      consumes:
      - application/json
      description: Updates a session by ID.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateSessionDTO'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update session
      tags:
      - session
  /client/{client_code}/session/{id}/messages:
    get:
      description: Lists the execution messages of a session, with optional ordering/pagination.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Order by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SessionMessagesPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List session messages
      tags:
      - session
  /client/{client_code}/session/{id}/run:
    post:
      consumes:
      - application/json
      description: Runs an agent within a session, streaming the execution via SSE.
      parameters:
      - description: Client code
        in: path
        name: client_code
        required: true
        type: string
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Run input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/RunSessionDTO'
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Run session
      tags:
      - session
...
  /llm:
    get:
      description: Lists configured LLMs, with optional pagination.
      parameters:
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LLMsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List LLMs
      tags:
      - llm
    post:
      consumes:
      - application/json
      description: Creates a new LLM configuration.
      parameters:
      - description: LLM data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateLLMDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/LLMResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Create LLM
      tags:
      - llm
  /llm/{id}:
    delete:
      description: Deletes an LLM by ID.
      parameters:
      - description: LLM ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete LLM
      tags:
      - llm
    get:
      description: Returns an LLM by ID.
      parameters:
      - description: LLM ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LLMResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Get LLM
      tags:
      - llm
    put:
      consumes:
      - application/json
      description: Updates an LLM by ID.
      parameters:
      - description: LLM ID
        in: path
        name: id
        required: true
        type: string
      - description: LLM data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateLLMDTO'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Update LLM
      tags:
      - llm
  /llm/drivers:
    get:
      description: Lists the available LLM provider drivers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DriverInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List LLM drivers
      tags:
      - llm
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"